saveCourses and saveAuthors moved from App to Courses
