+++ src/App.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { Switch, BrowserRouter as Router, Redirect } from 'react-router-dom';
 
 import Header from './components/Header/Header';
@@ -12,25 +12,11 @@ import UnauthenticatedRoute from './components/Routes/UnauthenticatedRoute';
 import { AppContext } from './AppContext';
 
 import './App.css';
-import { loadAuthors, loadCourses } from './services';
-import { useDispatch } from 'react-redux';
-import { saveCourses } from './store/courses/actionCreators';
-import { saveAuthors } from './store/authors/actionCreators';
 
 function App() {
 	const [isAuthenticated, setIsAuthenticated] = useState(
 		localStorage.getItem('token')
 	);
-	const dispatch = useDispatch();
-
-	useEffect(() => {
-		loadCourses().then((response) => {
-			dispatch(saveCourses(response.data.result));
-		});
-		loadAuthors().then((response) => {
-			dispatch(saveAuthors(response.data.result));
-		});
-	}, [dispatch]);
 
 	return (
 		<div className='App'>
